requirements.txt
absl-py==0.9.0
astroid==2.2.5
astunparse==1.6.3
attrs==19.3.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1588608214987/work
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
cloudpickle==1.3.0
colorama==0.4.1
cycler==0.10.0
Cython==0.29.20
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
-e git+https://github.com/openai/gym@809a8ad21ba66b2a17987005a04be3881808bf37#egg=gym
h5py==2.10.0
idna==2.9
importlib-metadata @ file:///D:/bld/importlib-metadata_1591451966421/work
ipykernel @ file:///D:/bld/ipykernel_1590023017759/work/dist/ipykernel-5.3.0-py3-none-any.whl
ipython @ file:///D:/bld/ipython_1590797082149/work
ipython-genutils==0.2.0
isort==4.3.21
jedi @ file:///D:/bld/jedi_1592619562975/work
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.4
jupyter-nbextensions-configurator==0.4.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.1
lxml @ file:///D:/bld/lxml_1589927701860/work
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib @ file:///C:/ci/matplotlib-base_1592401131947/work
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
MouseInfo==0.1.3
mss==5.1.0
nbconvert==5.6.1
nbformat==5.0.6
notebook @ file:///D:/bld/notebook_1588887315004/work
numpy==1.19.0
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1589925210001/work
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1590412252446/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1592500439797/work
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
pyglet==1.5.0
Pygments==2.6.1
pylint==2.3.1
PyMsgBox==1.0.8
pyparsing==2.4.7
pyperclip==1.8.0
PyRect==0.1.4
pyrsistent==0.16.0
PyScreeze==0.1.26
pystan==2.19.0.0
python-dateutil==2.8.1
PyTweening==1.0.3
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
Send2Trash==1.5.0
six==1.12.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow-gpu==2.2.0
tensorflow-gpu-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
tsfresh==0.12.0
typed-ast==1.4.0
urllib3==1.25.9
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1591600393557/work
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.11.2
zipp==3.1.0
